Added a new key to the config for location, added a test to check that dt datetime is in correct format

+++ OpenWeatherMapTest/OpenWeatherMap_Forecast_Service/Data_Handling/OpenWeatherMapForecastModel.cs
@@ -10,7 +10,7 @@ namespace OpenWeatherMapTest.OpenWeatherMap_Forecast_Service.Data_Handling
     {
         public class Forecast
         {
-            public int dt { get; set; }
+            public double dt { get; set; }
             public Main main { get; set; }
             public IList<Weather> weather { get; set; }
             public Clouds clouds { get; set; }
